requirements.txt
numpy
pandas
scikit-learn
tensorflow
requests
ccxt
matplotlib
python-dotenv
ta